+++ package.json
@@ -12,11 +12,13 @@
 		"@prisma/client": "^4.11.0",
 		"cors": "^2.8.5",
 		"dotenv": "^16.0.3",
-		"express": "^4.18.2"
+		"express": "^4.18.2",
+		"http-errors": "^2.0.0"
 	},
 	"devDependencies": {
 		"@types/cors": "^2.8.13",
 		"@types/express": "^4.17.17",
+		"@types/http-errors": "^2.0.1",
 		"@types/node": "^18.15.5",
 		"@typescript-eslint/eslint-plugin": "^5.56.0",
 		"@typescript-eslint/parser": "^5.56.0",
@@ -24,6 +26,7 @@
 		"eslint": "^8.36.0",
 		"nodemon": "^2.0.21",
 		"prisma": "^4.11.0",
+		"tasks-types": "^1.0.1",
 		"typescript": "^5.0.2"
 	}
 }
